% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{add_manifest_entity_column}
\alias{add_manifest_entity_column}
\title{Add Manifest Entity Column}
\usage{
add_manifest_entity_column(table, synapse_object)
}
\arguments{
\item{table}{A dataframe with column "manifest_id" that contains synapse
file ids for manifests.}

\item{synapse_object}{object returned by create_synapse_login().}
}
\description{
Maps synapse GET on the "manifest_id" and add the result entity on as
"manifest_entity" column.
}
