% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{add_manifest_path_column}
\alias{add_manifest_path_column}
\title{Add Manifest Path Column}
\usage{
add_manifest_path_column(table)
}
\arguments{
\item{table}{A dataframe with column "manifest_entity" that contains manifest
synapse file entities.}
}
\description{
Pulls the manifest path from the entities in the "manifest_entity" column,
and adds it as the "manifest_path" column.
}
