% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{add_manifest_tbl_column}
\alias{add_manifest_tbl_column}
\title{Add Manifest Table Column}
\usage{
add_manifest_tbl_column(table)
}
\arguments{
\item{table}{A dataframe with columns "manifest_path", "manifest_id" and
"manifest_entity"}
}
\description{
Returns a table with the manifest table as a column
}
