% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_object_from_table.R
\name{combine_component_tables}
\alias{combine_component_tables}
\title{Combine Component Tables}
\usage{
combine_component_tables(manifest_table_list, fileview)
}
\arguments{
\item{manifest_table_list}{A list of dataframes, with column "entityId"}

\item{fileview}{A dataframe with column "id"}
}
\description{
Formats each manifgest in list, combines them, and joins them to the fileview
}
