% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_object_from_table.R
\name{create_all_files_table}
\alias{create_all_files_table}
\title{Create All Files Table}
\usage{
create_all_files_table(fileview, manifest_list)
}
\arguments{
\item{fileview}{A dataframe with column "id"}

\item{manifest_list}{A list of manifests, each with a column "id"}
}
\description{
Filters the fileview table for ids that in any of the manifests
}
