% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_object_from_table.R
\name{create_data_object_from_table}
\alias{create_data_object_from_table}
\title{Create Data Object From Table}
\usage{
create_data_object_from_table(manifests_table, fileview)
}
\arguments{
\item{manifests_table}{A dataframe, with columns "manifest_tbl", and
"component_label}

\item{fileview}{A dataframe with column "id"}
}
\description{
Creates a named list with an item per component in the table. For each
component, all manifests are formated, combined together, and joined to the
fileview. Finally one last item named "All Files" is added to the list that
is the filewview with only files that are found in the manifests.
}
