% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projectlive_data_object.R
\name{create_fileview_column_list}
\alias{create_fileview_column_list}
\title{Create Fileview Column List}
\usage{
create_fileview_column_list(columns = c())
}
\arguments{
\item{columns}{A list of columns to include("id", "name" "createdOn" and
"projectId", always included.)}
}
\description{
Creates a list of columns that will be used in the SQL query to get the
filewview
}
