% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projectlive_data_object.R
\name{create_fileview_project_id_filter}
\alias{create_fileview_project_id_filter}
\title{Create Fileview Project ID Filter}
\usage{
create_fileview_project_id_filter(
  project_ids_to_keep = NULL,
  project_ids_to_remove = NULL
)
}
\arguments{
\item{project_ids_to_keep}{A list of project ids to include}

\item{project_ids_to_remove}{A list of project ids to not include,
only used if project_ids_to_keep is NULL}
}
\description{
Creates a SQL filter for removing projects that are unwanted in the
filewview such as test projects.
}
