% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_object_from_table.R
\name{create_manifest_list}
\alias{create_manifest_list}
\title{Create Manifest List}
\usage{
create_manifest_list(manifests_table, fileview)
}
\arguments{
\item{manifests_table}{A dataframe, with columns "manifest_tbl", and
"component_label}

\item{fileview}{A dataframe with column "id"}
}
\description{
Creates a named list of manifest tables per component from a manifest tables.
}
