% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{create_manifest_table}
\alias{create_manifest_table}
\title{Create Manifest Table}
\usage{
create_manifest_table(table)
}
\arguments{
\item{table}{A dataframe with columns "manifest_path", "manifest_id" and
"manifest_entity"}
}
\description{
Returns a table with the manifest table as a column
}
