% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{create_manifest_table_from_list}
\alias{create_manifest_table_from_list}
\title{Create Manifest Table From List}
\usage{
create_manifest_table_from_list(project_manifest_list, synapse_object)
}
\arguments{
\item{project_manifest_list}{A nested list in this form:
list(
 "syn10" =  list(
   list(
      list("syn1", "dataset1"),
      list("syn2", "synapse_storage_manifest.csv"),
      list("Component1", "Component1")
    ),
    list(
      list("syn3", "dataset1"),
      list("syn4", "synapse_storage_manifest.csv"),
      list("Component2", "Component2")
    )
  ),
  "syn10" = list(
    list(
      list("syn5", "dataset1"),
      list("syn6", "synapse_storage_manifest.csv"),
      list("Component3", "Component3")
    ),
    list(
      list("syn7", "dataset1"),
      list("syn8", "synapse_storage_manifest.csv"),
      list("Component4", "Component4")
    )
  )
)}

\item{synapse_object}{An object created by create_syn_obj()}
}
\description{
Create Manifest Table From List
}
