% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{create_manifest_table_row}
\alias{create_manifest_table_row}
\title{Create Manifest Table Row}
\usage{
create_manifest_table_row(manifest)
}
\arguments{
\item{manifest}{A nested list in this form:
list(
  list("syn1", "dataset1"),
  list("syn2", "synapse_storage_manifest.csv"),
  list("Component1", "Component1")
)}
}
\description{
Creates a one row tibble with the data from the manifest as
columns, any empty strings ('') are replaced with NAs.
}
