% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_functions.R
\name{create_syn_obj}
\alias{create_syn_obj}
\title{Create Synapse Object}
\usage{
create_syn_obj(synapse_auth_token)
}
\arguments{
\item{synapse_auth_token}{A personal access token created from Synapse}
}
\description{
Returns a synapse object that has been loged in with the token
}
