% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_functions.R
\name{create_syn_tbl_query}
\alias{create_syn_tbl_query}
\title{Create Synapse Table Query}
\usage{
create_syn_tbl_query(synapse_id, columns = NULL, filters = NULL)
}
\arguments{
\item{synapse_id}{An id for a Synapse table}

\item{columns}{A list of columns name in the table}

\item{filters}{A list of SQL filters such as "id = 'syn2'"}
}
\description{
Crates a SQL query for a synapse table
}
