% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{create_synapse_error_list}
\alias{create_synapse_error_list}
\title{Create Synapse Error List}
\usage{
create_synapse_error_list(table)
}
\arguments{
\item{table}{A dataframe with columns "manifest_path", "manifest_id" and
"manifest_entity"}
}
\description{
Returns a list of manifetss in the table without a file path
}
