% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_object_from_table.R
\name{format_manifest}
\alias{format_manifest}
\title{Format Manifest}
\usage{
format_manifest(manifest)
}
\arguments{
\item{manifest}{A dataframe}
}
\description{
Converts all columsn to strings
}
