% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projectlive_data_object.R
\name{get_fileview_tbl}
\alias{get_fileview_tbl}
\title{Get Fileview Table}
\usage{
get_fileview_tbl(
  fileview_id,
  synapse_object,
  project_ids_to_keep = NULL,
  project_ids_to_remove = NULL,
  columns = c()
)
}
\arguments{
\item{fileview_id}{The synapse ID of the fileview}

\item{synapse_object}{An object returned by create_synapse_login()}

\item{project_ids_to_keep}{A list of project ids to include}

\item{project_ids_to_remove}{A list of project ids to not include,
only used if project_ids_to_keep is NULL}

\item{columns}{A list of columns to include("id", "name" and
are "createdOn" always included.)}
}
\description{
Returns a dataframe of the Synspae fileview given.
}
