% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest_tbl_from_list.R
\name{get_path_from_synapse_entity}
\alias{get_path_from_synapse_entity}
\title{Get Path From Synapse Entity}
\usage{
get_path_from_synapse_entity(entity)
}
\arguments{
\item{entity}{A Synsape File entity.}
}
\description{
Gets the path form a Synpase File entity, or returns NA.
}
