% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_projectlive_data_object.R
\name{get_project_ids_from_fileview}
\alias{get_project_ids_from_fileview}
\title{Get Project IDs From Fileview}
\usage{
get_project_ids_from_fileview(fileview)
}
\arguments{
\item{fileview}{A tibble of the Synapse fileview}
}
\description{
Gets a list of project ids from the fileview, sorted and unique.
}
