% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_manifest_lists_from_project_ids.R
\name{get_project_manifest_lists}
\alias{get_project_manifest_lists}
\title{Get Project Manifest Lists}
\usage{
get_project_manifest_lists(project_ids, fileview_id, synapse_token)
}
\arguments{
\item{project_ids}{A list of synapse ID for the project to get the manifests
from.}

\item{fileview_id}{A synapse ID for the fileview to use.}

\item{synapse_token}{A synapse PAT with download priveledges.}
}
\description{
Gets the manifest synapse ids associated with a list of projects, and
separates the results into successes and errors.
}
