% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_manifest_lists_from_project_ids.R
\name{get_project_manifests}
\alias{get_project_manifests}
\title{Get Project Manifests}
\usage{
get_project_manifests(
  project_id,
  fileview_id,
  synapse_token,
  url = "http://localhost:3001/v1/storage/project/manifests"
)
}
\arguments{
\item{project_id}{A synapse ID for the project to get the manifests from.}

\item{fileview_id}{A synapse ID for the fileview to use.}

\item{synapse_token}{A synapse PAT with download priveledges.}

\item{url}{A URL to the schematic API.}
}
\description{
A call to the schamtic API, getting all manifest synapse ids for a project,
using thefileview(asset_view) for that project.
}
