% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_functions.R
\name{get_syn_tbl}
\alias{get_syn_tbl}
\title{Get Synapse Table}
\usage{
get_syn_tbl(synapse_id, syn, columns = NULL, filters = NULL)
}
\arguments{
\item{synapse_id}{An id for a Synapse table}

\item{syn}{An object returned by create_synapse_login().}

\item{columns}{A list of columns name in the table}

\item{filters}{A list of SQL filters such as "id = 'syn2'"}
}
\description{
Creates a SQL query for a Synapse table, runs it, reads the result file,
and returns a tibble.
}
