% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_project_manifest_lists_from_project_ids.R
\name{map_get_project_manifest_list_from_project_ids}
\alias{map_get_project_manifest_list_from_project_ids}
\title{Map Get Project Manifest List From Project IDs}
\usage{
map_get_project_manifest_list_from_project_ids(
  project_ids,
  fileview_id,
  synapse_token
)
}
\arguments{
\item{project_ids}{A list of synapse ID for the project to get the manifests
from.}

\item{fileview_id}{A synapse ID for the fileview to use.}

\item{synapse_token}{A synapse PAT with download priveledges.}
}
\description{
Attempts to get the manifest synapse ids associated with a list of projects.
}
