% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_functions.R
\name{map_synapse_get_entities}
\alias{map_synapse_get_entities}
\title{Map Synapse Get Entities}
\usage{
map_synapse_get_entities(synapse_ids, synapse_object)
}
\arguments{
\item{synapse_ids}{A list of synapse file ids}

\item{synapse_object}{An object returned by create_synapse_login().}
}
\description{
Map Synapse Get Entities
}
