% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_functions.R
\name{store_file_in_synapse}
\alias{store_file_in_synapse}
\title{Title}
\usage{
store_file_in_synapse(syn, file, parent_id, remove_file = TRUE)
}
\arguments{
\item{syn}{An object returned by create_synapse_login().}

\item{file}{The path to the file to be be strored}

\item{parent_id}{The synapse id of the folder where the file will be stored}

\item{remove_file}{T/F}
}
\description{
Title
}
