% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synapse_functions.R
\name{synapse_get}
\alias{synapse_get}
\title{Synapse Get}
\usage{
synapse_get(synapse_id, synapse_object)
}
\arguments{
\item{synapse_id}{A synapse id}

\item{synapse_object}{An object returned by create_synapse_login().}
}
\description{
Synapse Get
}
